+++ data-migrations/2025-02-27-0753-rename-plan-collaborators.sql
@@ -1,0 +1,2 @@
+
+RENAME TABLE planCollaborators TO projectCollaborators;